+++ e2e/test/exit-prematurely/stryker.conf.js
@@ -1,0 +1,19 @@
+module.exports = function (config) {
+  config.set({
+    mutate: [
+      'src/*.js'
+    ],
+    testFramework: 'mocha',
+    testRunner: 'mocha',
+    coverageAnalysis: 'off',
+    mutator: 'javascript',
+    transpilers: [
+      'babel'
+    ],
+    timeoutMS: 60000,
+    reporters: ['clear-text', 'html', 'event-recorder'],
+    maxConcurrentTestRunners: 2,
+    logLevel: 'info',
+    fileLogLevel: 'info'
+  });
+};