+++ e2e/test/karma-mocha/verify/verify.ts
@@ -1,13 +1,21 @@
-import { expectScoreResult } from '../../../helpers';
+import { expectMetricsResult, produceMetrics } from '../../../helpers';
 
 describe('Verify stryker has ran correctly', () => {
 
   it('should report correct score', async () => {
-    await expectScoreResult({
-      killed: 16,
-      mutationScore: 64,
-      noCoverage: 6,
-      survived: 3
+    await expectMetricsResult({
+      metrics: produceMetrics({
+        killed: 16,
+        mutationScore: 64,
+        mutationScoreBasedOnCoveredCode: 84.21,
+        noCoverage: 6,
+        survived: 3,
+        totalCovered: 19,
+        totalDetected: 16,
+        totalMutants: 25,
+        totalUndetected: 9,
+        totalValid: 25
+      })
     });
   });
 });