+++ e2e/test/mocha-mocha/verify/verify.ts
@@ -1,6 +1,6 @@
-import { expectScoreResult } from '../../../helpers';
-import * as fs from 'fs';
 import { expect } from 'chai';
+import * as fs from 'fs';
+import { expectMetricsResult, produceMetrics } from '../../../helpers';
 
 describe('Verify stryker has ran correctly', () => {
 
@@ -9,11 +9,18 @@ describe('Verify stryker has ran correctly', () => {
   }
 
   it('should report correct score', async () => {
-    await expectScoreResult({
-      killed: 16,
-      mutationScore: 64,
-      noCoverage: 0,
-      survived: 9
+    await expectMetricsResult({
+      metrics: produceMetrics({
+        killed: 16,
+        mutationScore: 64,
+        mutationScoreBasedOnCoveredCode: 64,
+        survived: 9,
+        totalCovered: 25,
+        totalDetected: 16,
+        totalMutants: 25,
+        totalUndetected: 9,
+        totalValid: 25
+      })
     });
   });