+++ e2e/test/vue-javascript/verify/verify.ts
@@ -1,19 +1,21 @@
-import { fsAsPromised } from '@stryker-mutator/util';
-import { expect } from 'chai';
-import * as path from 'path';
-import { ScoreResult } from '@stryker-mutator/api/report';
+import { expectMetricsResult, produceMetrics } from '../../../helpers';
 
 describe('After running stryker on VueJS project', () => {
   it('should report 25% mutation score', async () => {
-    const eventsDir = path.resolve(__dirname, '..', 'reports', 'mutation', 'events');
-    const allReportFiles = await fsAsPromised.readdir(eventsDir);
-    const scoreResultReportFile = allReportFiles.find(file => !!file.match(/.*onScoreCalculated.*/));
-    expect(scoreResultReportFile).ok;
-    const scoreResultContent = await fsAsPromised.readFile(path.resolve(eventsDir, scoreResultReportFile || ''), 'utf8');
-    const scoreResult = JSON.parse(scoreResultContent) as ScoreResult;
-    expect(scoreResult.killed).eq(4);
-    expect(scoreResult.survived).eq(12);
-
-    expect(scoreResult.mutationScore).to.equal(25);
+    await expectMetricsResult({
+      metrics: produceMetrics({
+        killed: 4,
+        mutationScore: 25,
+        mutationScoreBasedOnCoveredCode: 25,
+        runtimeErrors: 1,
+        survived: 12,
+        totalCovered: 16,
+        totalDetected: 4,
+        totalInvalid: 1,
+        totalMutants: 17,
+        totalUndetected: 12,
+        totalValid: 16
+      })
+    });
   });
 });