+++ packages/mocha-runner/src/MochaOptions.ts
@@ -1,5 +1,8 @@
-
-declare interface MochaOptions {
+/**
+ * This is a subset of all possible mocha options
+ * Only those that come in handy for the mocha runner plugin.
+ */
+export interface MochaOptions {
   require?: string[];
   opts?: string;
   config?: string;
@@ -8,6 +11,7 @@ declare interface MochaOptions {
   asyncOnly?: boolean;
   ui?: string;
   files?: string[] | string;
+  spec?: string[];
   grep?: RegExp;
   extension?: string[];
 }