+++ packages/mocha-runner/test/unit/MochaOptionsLoader.spec.ts
@@ -7,6 +7,7 @@ import sinon = require('sinon');
 import { testInjector } from '@stryker-mutator/test-helpers';
 import LibWrapper from '../../src/LibWrapper';
 import { mochaOptionsKey } from '../../src/utils';
+import { MochaOptions } from '../../src/MochaOptions';
 
 describe(MochaOptionsLoader.name, () => {
 
@@ -31,17 +32,18 @@ describe(MochaOptionsLoader.name, () => {
     it('should log about mocha >= 6', () => {
       sut.load(testInjector.options);
       expect(testInjector.logger.debug).calledWith(
-        'Mocha > 6 detected. Using mocha\'s `%s` to load mocha options', LibWrapper.loadOptions && LibWrapper.loadOptions.name
+        'Mocha >= 6 detected. Using mocha\'s `%s` to load mocha options', LibWrapper.loadOptions && LibWrapper.loadOptions.name
       );
     });
 
     it('should call `loadOptions` with serialized arguments', () => {
       testInjector.options[mochaOptionsKey] = {
-        baz: true,
-        foo: 'bar'
+        ['no-baz']: true,
+        foo: 'bar',
+        spec: ['helpers/*.js', 'test/*.js']
       };
       sut.load(testInjector.options);
-      expect(LibWrapper.loadOptions).calledWith(['--baz', 'true', '--foo', 'bar']);
+      expect(LibWrapper.loadOptions).calledWith(['--no-baz', '--foo', 'bar', '--spec', 'helpers/*.js,test/*.js']);
     });
 
     it('should filter out invalid options from the `loadOptions` result', () => {
@@ -58,6 +60,7 @@ describe(MochaOptionsLoader.name, () => {
       rawOptions.grep = 'quuz';
       rawOptions.exclude = 'corge';
       rawOptions.file = 'grault';
+      rawOptions.spec = ['test/**/*.js'];
 
       rawOptions.garply = 'waldo'; // this should be filtered out
       const result = sut.load(testInjector.options);
@@ -66,8 +69,10 @@ describe(MochaOptionsLoader.name, () => {
         extension: 'foo',
         file: 'grault',
         grep: 'quuz',
+        opts: './test/mocha.opts',
         override: true,
         require: 'bar',
+        spec: ['test/**/*.js'],
         timeout: 'baz',
         ['async-only']: 'qux',
         ui: 'quux',
@@ -86,6 +91,11 @@ describe(MochaOptionsLoader.name, () => {
         `Mocha: ${fnName}(['--foo','bar']) => {"baz":"qux"}`
       );
     });
+
+    it('should respect mocha\'s defaults', () => {
+      const options = sut.load(testInjector.options);
+      expect(options).deep.eq(defaultMochaOptions());
+    });
   });
 
   describe('with mocha < 6', () => {
@@ -101,6 +111,12 @@ describe(MochaOptionsLoader.name, () => {
       existsFileStub.returns(false);
       sut.load(config);
       expect(testInjector.logger.debug).calledWith('Mocha < 6 detected. Using custom logic to parse mocha options');
+    });
+
+    it('should log deprecated mocha version warning', async () => {
+      existsFileStub.returns(false);
+      sut.load(config);
+      expect(testInjector.logger.warn).calledWith('DEPRECATED: Mocha < 6 detected. Please upgrade to at least Mocha version 6.');
     });
 
     it('should load a mocha.opts file if specified', () => {
@@ -128,7 +144,7 @@ describe(MochaOptionsLoader.name, () => {
       readFileStub.returns('');
       sut.load(config);
       expect(testInjector.logger.info).calledWith(`Loading mochaOpts from "${path.resolve('test/mocha.opts')}"`);
-      expect(fs.readFileSync).calledWith(path.resolve('test/mocha.opts'));
+      expect(fs.readFileSync).calledWith(path.resolve('./test/mocha.opts'));
     });
 
     it('shouldn\'t load anything if mocha.opts = false', () => {
@@ -143,7 +159,7 @@ describe(MochaOptionsLoader.name, () => {
     it('should not load default mocha.opts file if not found', () => {
       existsFileStub.returns(false);
       const options = sut.load(config);
-      expect(options).deep.eq({});
+      expect(options).deep.eq(defaultMochaOptions());
       expect(testInjector.logger.debug).calledWith('No mocha opts file found, not loading additional mocha options (%s.opts was not defined).', 'mochaOptions');
     });
 
@@ -180,7 +196,7 @@ describe(MochaOptionsLoader.name, () => {
     itShouldLoadProperty('--grep', '/grep(this|that)/', { grep: /grep(this|that)/ });
     itShouldLoadProperty('--grep', 'grep(this|that)?', { grep: /grep(this|that)?/ });
 
-    it('should not override additional properties', () => {
+    it('should not override additional properties (except for defaults)', () => {
       readFileStub.returns(`
         -u qunit
         -t 2000
@@ -197,8 +213,10 @@ describe(MochaOptionsLoader.name, () => {
       const options = sut.load(config);
       expect(options).deep.equal({
         asyncOnly: false,
+        extension: ['js'],
         opts: 'path/to/opts/file',
         require: ['ts-node/register'],
+        spec: ['test'],
         timeout: 4000,
         ui: 'exports'
       });
@@ -213,7 +231,8 @@ describe(MochaOptionsLoader.name, () => {
         opts: 'some/mocha.opts/file',
       };
       const options = sut.load(config);
-      expect(options).deep.eq({ opts: 'some/mocha.opts/file' });
+      expect(options).not.have.property('reporter');
+      expect(options).not.have.property('ignore-leaks');
       expect(testInjector.logger.debug).calledWith('Ignoring option "--reporter" as it is not supported.');
       expect(testInjector.logger.debug).calledWith('Ignoring option "--ignore-leaks" as it is not supported.');
     });
@@ -227,8 +246,26 @@ describe(MochaOptionsLoader.name, () => {
         opts: 'some/mocha.opts/file',
       };
       const options = sut.load(config);
-      expect(options).deep.eq({ opts: 'some/mocha.opts/file', timeout: undefined, ui: undefined });
+      expect(options).deep.eq({
+        extension: ['js'],
+        opts: 'some/mocha.opts/file',
+        spec: [
+          'test'
+        ],
+        timeout: undefined,
+        ui: undefined
+      });
     });
   });
+
+  function defaultMochaOptions()  {
+    return {
+      extension: ['js'],
+      opts: './test/mocha.opts',
+      spec: ['test'],
+      timeout: 2000,
+      ui: 'bdd'
+    };
+  }
 
 });